fix: исправлены ошибки типизации и логики в коде
Amp-Thread-ID: https://ampcode.com/threads/T-019cec8f-3a9a-75ad-9870-9e344e744d72

+++ parser_2gis/tui_pytermgui/widgets/progress_bar.py
@@ -86,4 +86,4 @@ class ProgressBar:
         """Процент выполнения."""
         if self._total <= 0:
             return 0.0
-        return min(100.0, max(0.0, (self._completed / self._total) * 100))
+        return float(min(100.0, max(0.0, (self._completed / self._total) * 100)))
